refactor: strict linters
