+++ internal/domain/const.go
@@ -20,6 +20,6 @@ http_request_timeout: 10
 `
 )
 
-var (
-	IgnoreHosts = []string{"localhost", "example.com", "example.org", "example.net", "127.0.0.1"}
-)
+var IgnoreHosts = []string{ //nolint:gochecknoglobals
+	"localhost", "example.com", "example.org", "example.net", "127.0.0.1",
+}
