+++ internal/handler/init.go
@@ -9,8 +9,8 @@ import (
 	"github.com/suzuki-shunsuke/durl/internal/usecase"
 )
 
-// InitCommand is the sub command "init".
-var InitCommand = cli.Command{
+// initCommand is the sub command "init".
+var initCommand = cli.Command{ //nolint:gochecknoglobals
 	Name:   "init",
 	Usage:  "create a configuration file if it doesn't exist",
 	Action: initCmd,
