+++ internal/usecase/check_internal_test.go
@@ -46,11 +46,15 @@ func Test_logicCheck(t *testing.T) {
 	}{{
 		"normal", test.NewLogic(t, gomic.DoNothing), require.Nil,
 	}, {
-		"failed to get file paths", test.NewLogic(t, gomic.DoNothing).SetReturnGetFiles(nil, fmt.Errorf("failed to get file paths")), require.NotNil,
+		"failed to get file paths",
+		test.NewLogic(t, gomic.DoNothing).SetReturnGetFiles(nil, fmt.Errorf("failed to get file paths")), require.NotNil,
 	}, {
-		"failed to extract urls from files", test.NewLogic(t, gomic.DoNothing).SetReturnExtractURLsFromFiles(nil, fmt.Errorf("failed to extract urls from files")), require.NotNil,
+		"failed to extract urls from files",
+		test.NewLogic(t, gomic.DoNothing).SetReturnExtractURLsFromFiles(nil, fmt.Errorf("failed to extract urls from files")),
+		require.NotNil,
 	}}
 	for _, tt := range data {
+		tt := tt
 		t.Run(tt.title, func(t *testing.T) {
 			lgc := &logic{
 				logic: tt.mock,
@@ -98,13 +102,15 @@ func Test_logicCheckURLs(t *testing.T) {
 		test.NewLogic(t, gomic.DoNothing),
 		map[string]*strset.Set{
 			"http://example.com/foo": strset.New("foo.txt"),
-		}, require.Nil,
+		},
+		require.Nil,
 	}, {
 		"urls is empty",
 		test.NewLogic(t, gomic.DoNothing), nil, require.Nil,
 	}}
 	cfg := domain.Cfg{HTTPMethod: "head,get"}
 	for _, tt := range data {
+		tt := tt
 		t.Run(tt.title, func(t *testing.T) {
 			lgc := &logic{
 				logic: tt.mock,
@@ -142,6 +148,7 @@ func Test_logicCheckURLWithMethod(t *testing.T) {
 		true,
 	}}
 	for _, tt := range data {
+		tt := tt
 		t.Run(tt.title, func(t *testing.T) {
 			lgc := &logic{
 				client: tt.client,
@@ -187,6 +194,7 @@ func Test_logicCheckURL(t *testing.T) {
 		Timeout: domain.DefaultTimeout,
 	}
 	for _, tt := range data {
+		tt := tt
 		t.Run(tt.title, func(t *testing.T) {
 			lgc := &logic{
 				logic:  tt.mock,
@@ -213,15 +221,18 @@ func Test_logicExtractURLsFromFiles(t *testing.T) {
 			"foo.txt": {[]byte(`foo`), nil},
 			"bar.txt": {[]byte(`http://example.com`), nil},
 		}, require.Nil, map[string]*strset.Set{
-			"http://example.com": strset.New("bar.txt")},
+			"http://example.com": strset.New("bar.txt"),
+		},
 	}, {
 		"error", map[string]File{
 			"bar.txt": {[]byte(`http://example.com`), nil},
 			"foo.txt": {nil, fmt.Errorf("failed to read a file")},
 		}, require.Nil, map[string]*strset.Set{
-			"http://example.com": strset.New("bar.txt")},
+			"http://example.com": strset.New("bar.txt"),
+		},
 	}}
 	for _, tt := range data {
+		tt := tt
 		t.Run(tt.title, func(t *testing.T) {
 			fsys := newFsys(t, tt.files)
 			files := strset.New()
@@ -257,6 +268,7 @@ bar`), nil, require.Nil, strset.New(), "foo.txt",
 		"too long", []byte(strings.Repeat("X", 65536)), nil, require.NotNil, nil, "foo.txt",
 	}}
 	for _, tt := range data {
+		tt := tt
 		t.Run(tt.title, func(t *testing.T) {
 			var rc io.ReadCloser
 			if tt.buf != nil {
@@ -293,6 +305,7 @@ bar
 	}}
 	lgc := NewLogic(domain.Cfg{}, nil, nil)
 	for _, tt := range data {
+		tt := tt
 		t.Run(tt.title, func(t *testing.T) {
 			arr, err := lgc.GetFiles(bytes.NewBufferString(tt.in))
 			tt.checkErr(t, err)
